perf(search): lazy-mount Lexical editor on search bar focus
